haofuwu/backend/requirements.txt
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
SQLAlchemy>=1.4
passlib[bcrypt]>=1.7
python-jose>=3.3
python-multipart>=0.0.5
pydantic>=1.10